[feat] 텍스트 메시지 전송 기능 추가

+++ src/main/java/applesquare/moment/chat/dto/ChatMessageReadResponseDTO.java
@@ -23,6 +23,6 @@ public class ChatMessageReadResponseDTO {
     private String content;
     private List<String> fileUrls;
     private SharedPostReadResponseDTO sharedPost;
-    private int unreadCount;
+    private long unreadCount;
     private boolean isDeleted;
 }
